(lex) Move lexer helper functions to a util file
These functions all deal with characters, so `Characters` feels
reasonable.
